task_11/requirements.txt
openai
pydub
python-dotenv 